requirements.txt
fastapi
uvicorn
pydantic
pylint
pytest-cov
click
black
flake8
requests
python-jose[cryptography]
passlib[bcrypt]